# docs/swagger.yaml
basePath: /v1
definitions:
...
  attachment.insertProject:
    properties:
      comments:
        type: string
      content:
        type: string
      deliveries:
        items:
          type: string
        type: array
      due_date:
        type: string
      kind:
        type: string
      labels:
        items:
          properties:
            id:
              type: string
          required:
          - id
          type: object
        type: array
      person_id:
        type: string
      requirements:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - kind
    - requirements
    - title
    type: object
...
  attachment.updateProduct:
    properties:
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      name:
        type: string
      prices:
        items:
          properties:
            code:
              type: string
            description:
              type: string
            id:
              type: string
            status:
              type: boolean
            title:
              type: string
            value:
              type: number
          required:
          - code
          - description
          - value
          type: object
        type: array
      status:
        type: boolean
    required:
    - description
    - name
    type: object
...
  attachment.updateProject:
    properties:
      comments:
        type: string
      content:
        type: string
      deliveries:
        items:
          type: string
        type: array
      due_date:
        type: string
      kind:
        type: string
      labels:
        items:
          type: string
        type: array
      person_id:
        type: string
      requirements:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - kind
    - requirements
    - title
    type: object
...
  entities.Label:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  entities.Person:
    properties:
      addresses:
        items:
          $ref: '#/definitions/entities.PersonAddress'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      document:
        type: string
      emails:
        items:
          $ref: '#/definitions/entities.PersonEmail'
        type: array
      extras:
        items:
          $ref: '#/definitions/entities.PersonExtra'
        type: array
      id:
        type: string
      main_person_id:
        type: string
      name:
        type: string
      nickname:
        type: string
      phones:
        items:
          $ref: '#/definitions/entities.PersonPhone'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - document
    - name
    type: object
  entities.PersonAddress:
    properties:
      address:
        type: string
      city:
        type: string
      complement:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      is_main:
        type: boolean
      neighborhood:
        type: string
      number:
        type: string
      person_id:
        type: string
      state:
        type: string
      updated_at:
        type: string
      zip_code:
        type: string
    required:
    - address
    - city
    - country
    - neighborhood
    - person_id
    - state
    - zip_code
    type: object
  entities.PersonEmail:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: string
      is_main:
        type: boolean
      person_id:
        type: string
      updated_at:
        type: string
    required:
    - person_id
    type: object
  entities.PersonExtra:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      key:
        type: string
      person_id:
        type: string
      updated_at:
        type: string
      value:
        type: string
    required:
    - key
    - person_id
    - value
    type: object
  entities.PersonPhone:
    properties:
      area_code:
        type: integer
      country_code:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      is_main:
        type: boolean
      is_whatsapp:
        type: boolean
      person_id:
        type: string
      phone_number:
        type: integer
      updated_at:
        type: string
    required:
    - area_code
    - country_code
    - person_id
    - phone_number
    type: object
  entities.Product:
    properties:
      categories:
        items:
          $ref: '#/definitions/entities.ProductCategory'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      prices:
        items:
          $ref: '#/definitions/entities.ProductPrice'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - description
    - name
    type: object
  entities.ProductCategory:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/entities.Product'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - description
    - name
    type: object
  entities.ProductPrice:
    properties:
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      product_id:
        type: string
      status:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
      value:
        type: number
    required:
    - code
    - description
    - value
    type: object
  entities.Project:
    properties:
      comments:
        type: string
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deliveries:
        items:
          type: string
        type: array
      due_date:
        type: string
      id:
        type: string
      kind:
        type: string
      labels:
        items:
          $ref: '#/definitions/entities.Label'
        type: array
      person:
        $ref: '#/definitions/entities.Person'
      person_id:
        type: string
      repositories:
        items:
          $ref: '#/definitions/entities.Repositories'
        type: array
      requirements:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    required:
    - content
    - kind
    - requirements
    - title
    type: object
  entities.Repositories:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    required:
    - name
    - url
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  helpers.Error:
    properties:
      payload:
        properties:
          code:
            example: 500
            type: integer
          details:
            example: error details
            type: string
          message:
            example: error message
            type: string
        required:
        - code
        - details
        - message
        type: object
      success:
        example: false
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseMany-array_entities_Project:
    properties:
      pagination:
        $ref: '#/definitions/utils.Pagination'
      payload:
        items:
          $ref: '#/definitions/entities.Project'
        type: array
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseNone:
    properties:
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseOne-entities_Person:
    properties:
      payload:
        $ref: '#/definitions/entities.Person'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseOne-entities_Product:
    properties:
      payload:
        $ref: '#/definitions/entities.Product'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseOne-entities_Project:
    properties:
      payload:
        $ref: '#/definitions/entities.Project'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseOne-two_factor_OtpResponse:
    properties:
      payload:
        $ref: '#/definitions/two_factor.OtpResponse'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  two_factor.OtpResponse:
    properties:
      qr_code:
        type: string
      secret:
        type: string
    required:
    - qr_code
    - secret
    type: object
  utils.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      pages:
        items:
          type: integer
        type: array
      total_pages:
        type: integer
      total_registers:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: Welcome to the official documentation for the The Mug Codes manager.
    This guide provides comprehensive information to integrate and utilize our API
    effectively. It includes details on authentication, endpoints, rate limits, error
    handling, best practices, and support. Whether you're a beginner or an experienced
    developer, this documentation serves as your roadmap to leverage our services
    seamlessly.
  title: The Mug Codes
  version: "1.0"
paths:
...
  /auth/sign-out:
    delete:
      consumes:
      - application/json
      description: Revokes a new user access credentials.
      parameters:
      - description: Username
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Sign-out User
      tags:
      - Security - Auth
  /auth/two-factor:
    delete:
      consumes:
      - application/json
      description: Deactivates two factor for validate sign-in.
      parameters:
      - description: Username
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Deactivate Two Factor Authentication for User
      tags:
      - Security - Two Factor
    get:
      consumes:
      - application/json
      description: Generate two factor for validate sign-in.
      parameters:
      - description: Username
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ResponseOne-two_factor_OtpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Generate Two Factor Authentication for User
      tags:
      - Security - Two Factor
    post:
      consumes:
      - application/json
      description: Activates two factor for validate sign-in.
      parameters:
      - description: Two factor code
        in: query
        name: code
        required: true
        type: string
      - description: Username
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Activate Two Factor Authentication for User
      tags:
      - Security - Two Factor
...
  /products/{id}:
    delete:
      description: Removes a product by idfrom database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Delete a Product
      tags:
      - Products
    get:
      description: Get a product by idin database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show a Products
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates a product by idin database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.updateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Change a Product
      tags:
      - Products
...
  /projects:
    get:
      description: Get all projects from database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseMany-array_entities_Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show All Projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Creates a new project in database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.insertProject'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Create a Project
      tags:
      - Projects
  /projects/{id}:
    delete:
      description: Removes a project by idfrom database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Delete a Project
      tags:
      - Projects
    get:
      description: Get a project by idin database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Person'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show a Project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates a project by idin database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.updateProject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Change a Project
      tags:
      - Projects
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"